+++ blocklet/agent-runtime/api/src/routes/index.ts
@@ -2,32 +2,82 @@ import { AIGNEHubChatModel, loadModel } from "@aigne/aigne-hub";
 import { AIGNE } from "@aigne/core";
 import { loadAIGNEFile } from "@aigne/core/loader/index.js";
 import { AIGNEHTTPServer } from "@aigne/transport/http-server/server.js";
-import { Router } from "express";
+import { type NextFunction, type Request, type Response, Router } from "express";
+import path from "path";
+
+import logger from "../libs/logger.js";
 
 const router = Router();
+const AuthService = require("@blocklet/sdk/service/auth");
+
+const blocklet = new AuthService();
+const COMPONENT_CACHE_TTL = 1 * 60 * 1000; // 1 minute
+const componentCache = new Map();
+
+router.use(async (req: Request, res: Response, next: NextFunction) => {
+  const raw = String(req.headers["x-blocklet-component-id"] || "");
+  const componentId = raw.split("/").pop();
+
+  if (!componentId) {
+    return res.status(400).send("Agent Runtime: Bad Request");
+  }
+
+  const cachedChecker = componentCache.get(componentId);
+  if (!cachedChecker || Date.now() - cachedChecker.timestamp > COMPONENT_CACHE_TTL) {
+    const component = await blocklet.getComponent(componentId);
+    componentCache.set(componentId, {
+      component,
+      timestamp: Date.now(),
+    });
+  }
+
+  const cached = componentCache.get(componentId);
+  const { component } = cached;
+  if (!component) {
+    return res.status(404).send("Static Server: Component Not Found");
+  }
+
+  const env = component.environments.find((x: { key: string }) => x.key === "BLOCKLET_APP_DIR");
+  if (!env) {
+    return res.status(404).send("Static Server: Component Not Valid");
+  }
+
+  req.mainDir = component.meta.main
+    ? path.join(env.value, component.meta.main)
+    : path.resolve(env.value);
+  req.component = component;
+  if (process.env.DOCKER_HOST_SERVER_DIR && process.env.DOCKER_CONTAINER_SERVER_DIR) {
+    req.mainDir = req.mainDir.replace(
+      new RegExp(process.env.DOCKER_HOST_SERVER_DIR, "g"),
+      process.env.DOCKER_CONTAINER_SERVER_DIR,
+    );
+  }
+  logger.warn("serve static from", req.mainDir);
+  return next();
+});
 
 router.get("/health", (req, res) => {
   res.send(`component ${req.component.meta?.title ?? req.component.meta?.name} is running`);
 });
 
-router.post("/chat", async (req, res) => {
-  const { aigne } = await loadAIGNEFile(req.mainDir).catch(() => ({
-    aigne: null,
-  }));
-  const modelFromPath = `${aigne?.model?.provider ?? "openai"}/${aigne?.model?.name ?? "gpt-5-mini"}`;
-  const model = await AIGNEHubChatModel.load({ model: modelFromPath });
-  const engine = await AIGNE.load(req.mainDir, { model, loadModel });
-  const aigneServer = new AIGNEHTTPServer(engine);
-  await aigneServer.invoke(req, res, {
-    userContext: { userId: req.user?.did },
-  });
+router.get("/chat/agent", async (req, res) => {
+  const { aigne } = await loadAIGNEFile(req.mainDir).catch(() => ({ aigne: null }));
+  res.json({ data: aigne?.agents?.[0] });
 });
 
-router.get("/chat/agent", async (req, res) => {
-  const { aigne } = await loadAIGNEFile(req.mainDir).catch(() => ({
-    aigne: null,
-  }));
-  res.json({ data: aigne?.agents?.[0] });
+router.post("/chat", async (req, res) => {
+  console.log("chat=======================", req.mainDir);
+  try {
+    const { aigne } = await loadAIGNEFile(req.mainDir).catch(() => ({ aigne: null }));
+    const modelFromPath = `${aigne?.model?.provider ?? "openai"}/${aigne?.model?.name ?? "gpt-5-mini"}`;
+    const model = await AIGNEHubChatModel.load({ model: modelFromPath });
+    const engine = await AIGNE.load(req.mainDir, { model, loadModel });
+    const aigneServer = new AIGNEHTTPServer(engine);
+    await aigneServer.invoke(req, res, { userContext: { userId: req.user?.did } });
+  } catch (error) {
+    console.log("error=======================", error);
+    throw error;
+  }
 });
 
 export default router;
